requirements.txt
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
huggingface-hub>=0.19.0
pyyaml>=6.0
